Parse all files and store in CSV
